requirements.txt
streamlit==1.28.1
opencv-python==4.8.1.78
deepface==0.0.79
gtts==2.4.0
pygame==2.5.2
numpy==1.24.3
Pillow==10.0.1
tensorflow==2.13.0